Moves CSS files for toolbar
