+++ packages/styles/toolbar.css
@@ -1,3 +1,37 @@
+.toolbar {
+  position: relative;
+  overflow: none;
+  display: flex;
+  align-items: center;
+  justify-content: space-between;
+  background-color: rgba(255,255,255,0);
+  box-sizing: border-box;
+  box-shadow: none;
+  width: 100%;
+  height: 40px;
+  padding: 24px 0;
+
+  transition-property: height background-color box-shadow;
+  transition-duration: 500ms;
+  transition-timing-function: ease;
+}
+
+.toolbar > span {
+  display: flex;
+  justify-content: space-between;
+  width: 600px;
+}
+
+.toolbar.expanded {
+  height: 120px;
+  background-color: #FFF;
+  box-shadow: 0px 2px 8px rgba(41, 49, 58, 0.15);
+}
+
+.toolbar .toolbar-icon-flip {
+  transform: rotate(180deg);
+}
+
 .toolbar-item {
   display: flex;
   flex-direction: column;
